Type:fix
Description:when  choose application combox is selected and table result is empty ,detail info maybe have context but it supposed to be empty.

RootCause:Null

Solution:the function slot_tableItemClicked and sigDetailInfoSlot check the modelIndex twice,the
detail widget clean work excute in sigDetailInfoSlot,but slot_tableItemClicked check and return.